requirements.txt
playwright
beautifulsoup4
pandas
pandas-stubs
pyyaml
supabase
colorlog
types-PyYAML